sdl3-port

+++ sdlgui/button.h
@@ -80,13 +80,13 @@ public:
     void setButtonGroup(const std::vector<Button *> &buttonGroup) { mButtonGroup = buttonGroup; }
     const std::vector<Button *> &buttonGroup() const { return mButtonGroup; }
 
-    virtual Vector2i preferredSize(SDL_Renderer *ctx) const override;
-    virtual bool mouseButtonEvent(const Vector2i &p, int button, bool down, int modifiers) override;
+    virtual Vector2f preferredSize(SDL_Renderer *ctx) const override;
+    virtual bool mouseButtonEvent(const Vector2f &p, int button, bool down, int modifiers) override;
     virtual void draw(SDL_Renderer* renderer) override;
     virtual void drawBody(SDL_Renderer* renderer);
     virtual void drawBodyTemp(SDL_Renderer* renderer);
     virtual Color bodyColor();
-    virtual Vector2i getTextOffset() const;
+    virtual Vector2f getTextOffset() const;
 
     Button& withCallback(const std::function<void()> &callback) { setCallback( callback ); return *this; }
     Button& withFlags(int flags) { setFlags( flags); return *this; }
@@ -95,7 +95,7 @@ public:
     Button& withIcon(int icon) { setIcon( icon ); return *this; }
 
 protected:
-    virtual void renderBodyTexture(NVGcontext* &ctx, int &realw, int &realh);
+    virtual void renderBodyTexture(NVGcontext* &ctx, float &realw, float &realh);
 
     std::string mCaption;
     intptr_t mIcon;
